// frontend/packages/workflow/playground/src/node-registries/intent/types.ts
/*
 * Copyright 2025 ynet-dev Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {
  type InputValueVO,
  type NodeDataDTO,
  type OutputValueVO,
} from '@coze-workflow/base';

export type IntentsType = { name?: string; id?: string }[];

export interface FormData {
  inputs: {
    inputParameters: InputValueVO[];
    chatHistorySetting: {
      enableChatHistory?: boolean;
      chatHistoryRound?: number;
    };
  };
  model: { [k: string]: unknown };
  intents: IntentsType;
  quickIntents: IntentsType;
  intentMode: string;
  systemPrompt: string;
  nodeMeta: NodeDataDTO['nodeMeta'];
  outputs: OutputValueVO[];
}